#15: Refactoring: move callback to textui
